+++ src/main/java/lk/ijse/company/model/tm/OrderTm.java
@@ -15,5 +15,5 @@ public class OrderTm {
     private String orderId;
     private String customerId;
     private LocalDate orderDate;
-    private BigDecimal orderTotal;
+    private double orderTotal;
 }
